Add experimental parser.
It was also necessary to introduce a concrete syntax because, for
example, it's difficult to bind variables while parsing, so we defer
that work to the translation from concrete to abstract.

The parser is stateless, based on Parsec, and using a lot of Parsec
niceties, like the bult in Haskell lexer.

One compromise to note: We adopt ocaml-style arithmetic operators, like
`+.`, `*.`, etc, since we can't yet infer the type of an operator (and
we don't have, e.g. typeclasses! (or defaulting)).